+++ examples/basemaps/openfreemap.py
@@ -1,0 +1,5 @@
+from maplibre import Map, MapOptions
+from maplibre.basemaps import OpenFreeMap
+
+m = Map(MapOptions(style=OpenFreeMap.BRIGHT))
+m.save(preview=True)