+++ maplibre/ipywidget.py
@@ -1,18 +1,29 @@
 from __future__ import annotations
 
+import logging
 from os.path import join
 from pathlib import Path
 
-import traitlets
-from anywidget import AnyWidget
+logging.basicConfig()
+logger = logging.getLogger(__name__)
 
-from .map import Map, MapOptions
+
+try:
+    import traitlets
+    from anywidget import AnyWidget
+except ImportError as e:
+    # traitlets = None
+    # AnyWidget = None
+    logger.error(e)
+    logger.error("Please install 'maplibre[anywidget]' if you want to use maplibre in 'marimo' or 'jupyter' notebooks.")
+
+from maplibre.map import Map, MapOptions
 
 
 class MapWidget(AnyWidget, Map):
     """MapWidget
 
-    Use this class to display and update maps in Jupyter Notebooks.
+    Use this class to display and update maps in Marimo or Jupyter Notebooks.
 
     See `maplibre.Map` for available methods.
 
@@ -24,10 +35,8 @@ class MapWidget(AnyWidget, Map):
     """
 
     _esm = join(Path(__file__).parent, "srcjs", "ipywidget.js")
-    # _css = join(Path(__file__).parent, "srcjs", "maplibre-gl.css")
     _css = join(Path(__file__).parent, "srcjs", "ipywidget.css")
     _use_message_queue = True
-    # _rendered = traitlets.Bool(False, config=True).tag(sync=True)
     _rendered = traitlets.Bool(False).tag(config=True).tag(sync=True)
     map_options = traitlets.Dict().tag(sync=True)
     calls = traitlets.List().tag(sync=True)
@@ -40,6 +49,9 @@ class MapWidget(AnyWidget, Map):
     # Interactions MapboxDraw plugin
     draw_features_selected = traitlets.List().tag(sync=True)
     draw_feature_collection_all = traitlets.Dict().tag(sync=True)
+    draw_features_created = traitlets.List().tag(sync=True)
+    draw_features_updated = traitlets.List().tag(sync=True)
+    draw_features_deleted = traitlets.List().tag(sync=True)
 
     def __init__(
         self,
@@ -53,12 +65,6 @@ class MapWidget(AnyWidget, Map):
         self.calls = []
         AnyWidget.__init__(self, height=height, **kwargs)
         Map.__init__(self, map_options, sources, layers, controls, **kwargs)
-
-    """
-    @traitlets.default("height")
-    def _default_height(self):
-        return "400px"
-    """
 
     @traitlets.validate("height")
     def _validate_height(self, proposal):